Update index.php
Changed image load parameter random to file date and time to reduce network load.
Parameter is used to enable detection of change of file in case file is updated. If parameter is used browser uses cached image. 
Earlier when random parameter was used, image was downloaded every time. Now it is downloaded when file change.
Browser seem to require couple cycles before it detects the file change and update the image in cache.